jubilee_streamlit/requirements.txt.txt
streamlit
gspread
oauth2client
google-api-python-client
pandas
Pillow
altair
xlsxwriter